Add ingest source fields and expand CI test coverage
- Add `ingest_source_kind` and `ingest_source_id` columns to event
  inserts across lookup and query integration tests to support
  richer event provenance tracking.
- Add `TestNativeMultivalueAgainstClickHouse` and
  `TestCanonicalDecimalPayloadTextSQLAgainstClickHouse` to the CI
  test suite with a dedicated runtime contracts step.
- Replace custom `stringPointer` helper with the `new` builtin in
  lookup tests.
- Add `ec_param_enc:named_curve` option to openssl key generation
  in deploy scripts to use named curve encoding for EC keys.

+++ internal/control/migrations.go
@@ -68,7 +68,7 @@ func (db *DB) VerifyCurrentMigrations(
 	if err != nil {
 		return MigrationIdentity{}, err
 	}
-	if appliedCount != uint32(len(loaded)) {
+	if uint64(appliedCount) != uint64(len(loaded)) {
 		appliedVersion := uint32(0)
 		if appliedCount > 0 {
 			appliedVersion = loaded[appliedCount-1].version
